Rename package `domain` to `core`
