+++ src/main/java/ua/nykyforov/service/question/SimpleQuestionService.java
@@ -3,8 +3,9 @@ package ua.nykyforov.service.question;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
-import ua.nykyforov.dao.QuestionDAO;
-import ua.nykyforov.service.quiz.domain.model.QuizQuestion;
+import ua.nykyforov.service.quiz.core.application.QuestionService;
+import ua.nykyforov.service.quiz.core.business.QuestionDAO;
+import ua.nykyforov.service.quiz.core.model.QuizQuestion;
 
 import java.util.Collection;
